Update story-of-a-tensor-torch.py
Warning: I'm a rookie on python and torch, so I've modified the source on VS Code using the "hello-ai-world-torch.py" file as a guide and many requests to GitHub Copilot selecting portion of the original code and asking "Do the same with Torch" 😄
After some iteration I made it working. I wish I coud use a Mac to check if my version returns equivalent output, especially regarding output of weight and bias parameters.